control: refactored ArmController base class
